requirements.txt
#
# This file is autogenerated by pip-compile with Python 3.12
# by the following command:
#
#    pip-compile
#
annotated-types==0.7.0
    # via pydantic
asgiref==3.8.1
    # via
    #   django
    #   django-stubs
django==5.0.6
    # via
    #   -r requirements.in
    #   django-ninja
    #   django-stubs
    #   django-stubs-ext
django-ninja==1.1.0
    # via -r requirements.in
django-stubs==5.0.2
    # via -r requirements.in
django-stubs-ext==5.0.2
    # via django-stubs
environs==11.0.0
    # via -r requirements.in
marshmallow==3.21.2
    # via environs
mypy==1.10.0
    # via -r requirements.in
mypy-extensions==1.0.0
    # via mypy
packaging==24.0
    # via marshmallow
pydantic==2.7.2
    # via django-ninja
pydantic-core==2.18.3
    # via pydantic
python-dotenv==1.0.1
    # via environs
ruff==0.4.7
    # via -r requirements.in
sqlparse==0.5.0
    # via django
types-pyyaml==6.0.12.20240311
    # via django-stubs
typing-extensions==4.12.1
    # via
    #   django-stubs
    #   django-stubs-ext
    #   mypy
    #   pydantic
    #   pydantic-core
